fix: fixed plugin for hyper 2.0.
Using CSS custom properties for colors/units variables was breaking
the whole rendering engine for some reasons; the solution was to simply
insert plain CSS values where needed.

For the plugin to work 100% correctly a backgroundColor property must be
set to an almost transparent value in the consumer .hyper.js file, or the
canvases in the center will have a solid color and the result is not at
all pleasant.